+++ persistent/Database/Persist/Types/Base.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE LambdaCase #-}
 {-# OPTIONS_GHC -fno-warn-deprecations #-} -- usage of Error typeclass
 module Database.Persist.Types.Base where
@@ -15,7 +16,11 @@ import Data.Char (isSpace)
 import qualified Data.HashMap.Strict as HM
 import Data.Int (Int64)
 import Data.Map (Map)
-import Data.Maybe
+import Data.Maybe ( isNothing )
+#if !MIN_VERSION_base(4,11,0)
+-- This can be removed when GHC < 8.2.2 isn't supported anymore
+import Data.Semigroup ((<>))
+#endif
 import qualified Data.Scientific
 import Data.Text (Text, pack)
 import qualified Data.Text as T
@@ -107,13 +112,36 @@ data WhyNullable = ByMaybeAttr
                  | ByNullableAttr
                   deriving (Eq, Show)
 
+-- | Convenience operations for working with '-NameDB' types.
+--
+-- @since 2.12.0.0
+class DatabaseName a where
+  escapeWith :: (Text -> str) -> (a -> str)
+
+-- | An 'EntityNameDB' represents the datastore-side name that @persistent@
+-- will use for an entity.
+--
+-- @since 2.12.0.0
+newtype EntityNameDB = EntityNameDB { unEntityNameDB :: Text }
+  deriving (Show, Eq, Read, Ord)
+
+instance DatabaseName EntityNameDB where
+  escapeWith f (EntityNameDB n) = f n
+
+-- | An 'EntityNameHS' represents the Haskell-side name that @persistent@
+-- will use for an entity.
+--
+-- @since 2.12.0.0
+newtype EntityNameHS = EntityNameHS { unEntityNameHS :: Text }
+  deriving (Show, Eq, Read, Ord)
+
 -- | An 'EntityDef' represents the information that @persistent@ knows
 -- about an Entity. It uses this information to generate the Haskell
 -- datatype, the SQL migrations, and other relevant conversions.
 data EntityDef = EntityDef
-    { entityHaskell :: !HaskellName
+    { entityHaskell :: !EntityNameHS
     -- ^ The name of the entity as Haskell understands it.
-    , entityDB      :: !DBName
+    , entityDB      :: !EntityNameDB
     -- ^ The name of the database table corresponding to the entity.
     , entityId      :: !FieldDef
     -- ^ The entity's primary key or identifier.
@@ -144,7 +172,7 @@ data EntityDef = EntityDef
 
 entitiesPrimary :: EntityDef -> Maybe [FieldDef]
 entitiesPrimary t = case fieldReference primaryField of
-    CompositeRef c -> Just $ (compositeFields c)
+    CompositeRef c -> Just $ compositeFields c
     ForeignRef _ _ -> Just [primaryField]
     _ -> Nothing
   where
@@ -156,9 +184,8 @@ entityPrimary t = case fieldReference (entityId t) of
     _ -> Nothing
 
 entityKeyFields :: EntityDef -> [FieldDef]
-entityKeyFields ent = case entityPrimary ent of
-    Nothing   -> [entityId ent]
-    Just pdef -> compositeFields pdef
+entityKeyFields ent =
+    maybe [entityId ent] compositeFields $ entityPrimary ent
 
 keyAndEntityFields :: EntityDef -> [FieldDef]
 keyAndEntityFields ent =
@@ -168,11 +195,6 @@ keyAndEntityFields ent =
 
 
 type ExtraLine = [Text]
-
-newtype HaskellName = HaskellName { unHaskellName :: Text }
-    deriving (Show, Eq, Read, Ord)
-newtype DBName = DBName { unDBName :: Text }
-    deriving (Show, Eq, Read, Ord)
 
 type Attr = Text
 
@@ -238,17 +260,35 @@ data FieldType
     | FTList FieldType
   deriving (Show, Eq, Read, Ord)
 
--- | A 'FieldDef' represents the information that @persistent@ knows about
+-- | An 'EntityNameDB' represents the datastore-side name that @persistent@
+-- will use for an entity.
+--
+-- @since 2.12.0.0
+newtype FieldNameDB = FieldNameDB { unFieldNameDB :: Text }
+  deriving (Show, Eq, Read, Ord)
+
+-- | @since 2.12.0.0
+instance DatabaseName FieldNameDB where
+  escapeWith f (FieldNameDB n) = f n
+
+-- | A 'FieldNameHS' represents the Haskell-side name that @persistent@
+-- will use for a field.
+--
+-- @since 2.12.0.0
+newtype FieldNameHS = FieldNameHS { unFieldNameHS :: Text }
+  deriving (Show, Eq, Read, Ord)
+
+-- | A 'FieldDef' represents the inormation that @persistent@ knows about
 -- a field of a datatype. This includes information used to parse the field
 -- out of the database and what the field corresponds to.
 data FieldDef = FieldDef
-    { fieldHaskell   :: !HaskellName
+    { fieldHaskell   :: !FieldNameHS
     -- ^ The name of the field. Note that this does not corresponds to the
     -- record labels generated for the particular entity - record labels
     -- are generated with the type name prefixed to the field, so
-    -- a 'FieldDef' that contains a @'HaskellName' "name"@ for a type
+    -- a 'FieldDef' that contains a @'FieldNameHS' "name"@ for a type
     -- @User@ will have a record field @userName@.
-    , fieldDB        :: !DBName
+    , fieldDB        :: !FieldNameDB
     -- ^ The name of the field in the database. For SQL databases, this
     -- corresponds to the column name.
     , fieldType      :: !FieldType
@@ -290,8 +330,8 @@ isFieldNotGenerated = isNothing . fieldGenerated
 -- 2) single field
 -- 3) embedded
 data ReferenceDef = NoReference
-                  | ForeignRef !HaskellName !FieldType
-                    -- ^ A ForeignRef has a late binding to the EntityDef it references via HaskellName and has the Haskell type of the foreign key in the form of FieldType
+                  | ForeignRef !EntityNameHS !FieldType
+                    -- ^ A ForeignRef has a late binding to the EntityDef it references via name and has the Haskell type of the foreign key in the form of FieldType
                   | EmbedRef EmbedEntityDef
                   | CompositeRef CompositeDef
                   | SelfReference
@@ -302,7 +342,7 @@ data ReferenceDef = NoReference
 -- But it is only used for fieldReference
 -- so it only has data needed for embedding
 data EmbedEntityDef = EmbedEntityDef
-    { embeddedHaskell :: !HaskellName
+    { embeddedHaskell :: !EntityNameHS
     , embeddedFields  :: ![EmbedFieldDef]
     } deriving (Show, Eq, Read, Ord)
 
@@ -310,9 +350,9 @@ data EmbedEntityDef = EmbedEntityDef
 -- But it is only used for embeddedFields
 -- so it only has data needed for embedding
 data EmbedFieldDef = EmbedFieldDef
-    { emFieldDB       :: !DBName
+    { emFieldDB    :: !FieldNameDB
     , emFieldEmbed :: Maybe EmbedEntityDef
-    , emFieldCycle :: Maybe HaskellName
+    , emFieldCycle :: Maybe EntityNameHS
     -- ^ 'emFieldEmbed' can create a cycle (issue #311)
     -- when a cycle is detected, 'emFieldEmbed' will be Nothing
     -- and 'emFieldCycle' will be Just
@@ -338,6 +378,24 @@ toEmbedEntityDef ent = embDef
                         _ -> Nothing
                     }
 
+-- | A 'ConstraintNameDB' represents the datastore-side name that @persistent@
+-- will use for a constraint.
+--
+-- @since 2.12.0.0
+newtype ConstraintNameDB = ConstraintNameDB { unConstraintNameDB :: Text }
+  deriving (Show, Eq, Read, Ord)
+
+-- | @since 2.12.0.0
+instance DatabaseName ConstraintNameDB where
+  escapeWith f (ConstraintNameDB n) = f n
+
+-- | An 'ConstraintNameHS' represents the Haskell-side name that @persistent@
+-- will use for a constraint.
+--
+-- @since 2.12.0.0
+newtype ConstraintNameHS = ConstraintNameHS { unConstraintNameHS :: Text }
+  deriving (Show, Eq, Read, Ord)
+
 -- Type for storing the Uniqueness constraint in the Schema.
 -- Assume you have the following schema with a uniqueness
 -- constraint:
@@ -347,13 +405,13 @@ toEmbedEntityDef ent = embDef
 --   UniqueAge age
 --
 -- This will be represented as:
--- UniqueDef (HaskellName (packPTH "UniqueAge"))
--- (DBName (packPTH "unique_age")) [(HaskellName (packPTH "age"), DBName (packPTH "age"))] []
+-- UniqueDef (ConstraintNameHS (packPTH "UniqueAge"))
+-- (ConstraintNameDB (packPTH "unique_age")) [(FieldNameHS (packPTH "age"), FieldNameDB (packPTH "age"))] []
 --
 data UniqueDef = UniqueDef
-    { uniqueHaskell :: !HaskellName
-    , uniqueDBName  :: !DBName
-    , uniqueFields  :: ![(HaskellName, DBName)]
+    { uniqueHaskell :: !ConstraintNameHS
+    , uniqueDBName  :: !ConstraintNameDB
+    , uniqueFields  :: ![(FieldNameHS, FieldNameDB)]
     , uniqueAttrs   :: ![Attr]
     }
     deriving (Show, Eq, Read, Ord)
@@ -366,13 +424,13 @@ data CompositeDef = CompositeDef
 
 -- | Used instead of FieldDef
 -- to generate a smaller amount of code
-type ForeignFieldDef = (HaskellName, DBName)
+type ForeignFieldDef = (FieldNameHS, FieldNameDB)
 
 data ForeignDef = ForeignDef
-    { foreignRefTableHaskell       :: !HaskellName
-    , foreignRefTableDBName        :: !DBName
-    , foreignConstraintNameHaskell :: !HaskellName
-    , foreignConstraintNameDBName  :: !DBName
+    { foreignRefTableHaskell       :: !EntityNameHS
+    , foreignRefTableDBName        :: !EntityNameDB
+    , foreignConstraintNameHaskell :: !ConstraintNameHS
+    , foreignConstraintNameDBName  :: !ConstraintNameDB
     , foreignFieldCascade          :: !FieldCascade
     -- ^ Determine how the field will cascade on updates and deletions.
     --
@@ -586,12 +644,14 @@ instance A.FromJSON PersistValue where
             Just ('s', t) -> return $ PersistText t
             Just ('b', t) -> either (\_ -> fail "Invalid base64") (return . PersistByteString)
                            $ B64.decode $ TE.encodeUtf8 t
-            Just ('t', t) -> fmap PersistTimeOfDay $ readMay t
-            Just ('u', t) -> fmap PersistUTCTime $ readMay t
-            Just ('d', t) -> fmap PersistDay $ readMay t
-            Just ('r', t) -> fmap PersistRational $ readMay t
-            Just ('o', t) -> maybe (fail "Invalid base64") (return . PersistObjectId) $
-                              fmap (i2bs (8 * 12) . fst) $ headMay $ readHex $ T.unpack t
+            Just ('t', t) -> PersistTimeOfDay <$> readMay t
+            Just ('u', t) -> PersistUTCTime <$> readMay t
+            Just ('d', t) -> PersistDay <$> readMay t
+            Just ('r', t) -> PersistRational <$> readMay t
+            Just ('o', t) -> maybe
+                (fail "Invalid base64")
+                (return . PersistObjectId . i2bs (8 * 12) . fst)
+                $ headMay $ readHex $ T.unpack t
             Just (c, _) -> fail $ "Unknown prefix: " ++ [c]
       where
         headMay []    = Nothing
@@ -613,12 +673,12 @@ instance A.FromJSON PersistValue where
             then PersistInt64 $ floor n
             else PersistDouble $ fromRational $ toRational n
     parseJSON (A.Bool b) = return $ PersistBool b
-    parseJSON A.Null = return $ PersistNull
+    parseJSON A.Null = return PersistNull
     parseJSON (A.Array a) = fmap PersistList (mapM A.parseJSON $ V.toList a)
     parseJSON (A.Object o) =
         fmap PersistMap $ mapM go $ HM.toList o
       where
-        go (k, v) = fmap ((,) k) $ A.parseJSON v
+        go (k, v) = (,) k <$> A.parseJSON v
 
 -- | A SQL data type. Naming attempts to reflect the underlying Haskell
 -- datatypes, eg SqlString instead of SqlVarchar. Different SQL databases may